Adding my interactions, title screens and score
I added my starting and ending screens as well as my eating or zap functions so the Jumper would die if it touched the Ufos or the Pillars. I gave the Jumper the ability to shoot a bullet, which would kill the aliens. The score of many you kill is tracked and displayed.